experimento-autorizar-actores/autenticador/requirements.txt
Flask==3.*
PyJWT[crypto]==2.*
cryptography==42.*
gunicorn==21.*
requests==2.32.3